changes the notice style
