+++ client/src/containers/App.css
@@ -1,0 +1,7 @@
+.notice {
+  text-align: center;
+}
+
+.notice-text {
+  font-size: 4em;
+}
