+++ client/src/containers/App.js
@@ -1,129 +1,25 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
 
-import { withStyles } from 'material-ui/styles';
-import Paper from 'material-ui/Paper';
-import Grid from 'material-ui/Grid';
+import './App.css';
+import Dashboard from '../components/Dashboard';
+import SimpleAppBar from '../components/AppBar';
 
-import PriceByTravelDurationScatterPlot from './PriceByTravelDurationScatterPlot';
-import ShortestDistanceMap from './ShortestDistanceMap';
-import ImageGrid from './ImageGrid';
-import AppBar from '../components/AppBar';
-
-const spacing = 16;
-const style = {
-  root: {
-    flexGrow: 1
-  },
-  paper: {
-    padding: 10,
-    margin: 10,
-    textAlign: 'center'
-  }
+const App = props => {
+  const { displayStatic, message } = props;
+  return (
+    <div>
+      <SimpleAppBar />
+      {displayStatic ? (
+        <div className="notice">
+          <p className="notice-text">{message}</p>
+        </div>
+      ) : (
+        <Dashboard />
+      )}
+    </div>
+  );
 };
-const styles = theme => style;
-
-/**
- * Checkout the answer by speckledcarp on how to handle window resizing in React
- * https://stackoverflow.com/a/42141641
- */
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { width: 0, height: 0 };
-    this.updateWindowDimensions = this.updateWindowDimensions.bind(this);
-    this.getComponentWidth = this.getComponentWidth.bind(this);
-    this.getComponentMaxHeight = this.getComponentMaxHeight.bind(this);
-  }
-
-  componentDidMount() {
-    this.updateWindowDimensions();
-    window.addEventListener('resize', this.updateWindowDimensions);
-  }
-
-  componentWillUnmount() {
-    window.removeEventListener('resize', this.updateWindowDimensions);
-  }
-
-  updateWindowDimensions() {
-    this.setState({ width: window.innerWidth, height: window.innerHeight });
-  }
-
-  getComponentWidth() {
-    let width;
-    if (this.state.width < 960) {
-      const whiteSpaceWidth = (style.paper.margin + style.paper.padding) * 2;
-      width = this.state.width - whiteSpaceWidth;
-    } else {
-      const whiteSpaceWidth =
-        spacing + (style.paper.margin + style.paper.padding) * 4;
-      width = (this.state.width - whiteSpaceWidth) / 2;
-    }
-    return width < 0 ? 0 : width;
-  }
-
-  getComponentMaxHeight() {
-    let height = this.state.height - 200;
-    return Math.max(height / 2, 280);
-  }
-
-  render() {
-    const { classes, displayStatic, message } = this.props;
-    const width = this.getComponentWidth();
-    const maxHeight = this.getComponentMaxHeight();
-
-    const wrapInPaper = comp => {
-      return (
-        <Grid item xs={12}>
-          <Paper
-            className={classes.paper}
-            style={
-              // https://stackoverflow.com/questions/8865458/how-do-i-vertically-center-text-with-css
-              displayStatic
-                ? {
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                    height: 400
-                  }
-                : {}
-            }
-          >
-            {displayStatic ? message : comp}
-          </Paper>
-        </Grid>
-      );
-    };
-
-    return (
-      <div className={classes.root}>
-        <Grid container spacing={spacing}>
-          <Grid item xs={12}>
-            <AppBar />
-          </Grid>
-          <Grid item xs={12} md={6}>
-            <Grid container spacing={spacing}>
-              {wrapInPaper(
-                <PriceByTravelDurationScatterPlot
-                  width={width}
-                  maxHeight={maxHeight}
-                />
-              )}
-              {wrapInPaper(<ImageGrid width={width} maxHeight={maxHeight} />)}
-            </Grid>
-          </Grid>
-          <Grid item xs={12} md={6}>
-            <Grid container spacing={spacing}>
-              {wrapInPaper(
-                <ShortestDistanceMap width={width} maxHeight={maxHeight} />
-              )}
-            </Grid>
-          </Grid>
-        </Grid>
-      </div>
-    );
-  }
-}
 
 const mapStateToProps = (state, ownProps) => {
   const isMissingData = state.data.length === 0;
@@ -135,4 +31,4 @@ const mapStateToProps = (state, ownProps) => {
   };
 };
 
-export default connect(mapStateToProps, () => ({}))(withStyles(styles)(App));
+export default connect(mapStateToProps, () => ({}))(App);
